requirements.txt
asgiref==3.2.7
click==7.1.2
Django==3.0.6
djangorestframework==3.11.0
joblib==0.15.1
nltk==3.5
numpy==1.18.4
pandas==1.0.3
python-dateutil==2.8.1
pytz==2020.1
regex==2020.5.14
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
sqlparse==0.3.1
threadpoolctl==2.0.0
tqdm==4.46.0
